[dingo-calcite] Add float that supports precision

+++ dingo-exec/src/main/java/io/dingodb/exec/type/converter/ExprConverter.java
@@ -33,8 +33,8 @@ public class ExprConverter implements DataConverter {
     }
 
     @Override
-    public Double convertFloatFrom(@NonNull Object value) {
-        return ((Number) value).doubleValue();
+    public Float convertFloatFrom(@NonNull Object value) {
+        return ((Number) value).floatValue();
     }
 
     @Override
